Extend IS [NOT] NULL predicate proofs for arrays
For the purposes of predicate proof testing we limit ScalarArrayOpExpr
decomposition to arrays with <= MAX_SAOP_ARRAY_SIZE items (currently
100 items). Unfortunately this means that we miss opportunities for
proofs with both large and non-constant array expressions.

It's trivially shown that a scalar array op with a known non-empty array
is strict so long its operator is strict; similarly ops with constant
NULL arrays always result in NULL and are therefore strict.

If we don't know an array is NULL or non-empty, things are a bit more
complicated. For ANY ops an empty array results in FALSE, so we can
continue to derive  proofs since in the base case we only care about
expressions that result in TRUE for proof correctness.

For ALL ops we can't make any assumptions: if the array isn't known to
be empty or NULL we have to pass on proving [NOT] NULL.

One other exception: once we've recursed through a strict function we
have to require strictness (rather than our relaxed "can be FALSE for
ANY ops" definition in the base case) since a strict function may return
TRUE at arbitrary points.

One of the main benefits of this change is that the planner may now use
partial indexes of the form "WHERE foo IS NOT NULL" when the query's
WHERE clause involves a scalar array op like "foo IN (1,2,...101)" or
"foo != ALL(ARRAY[1,2,...101])".