Correct pg_rewind explanation
The pg_rewind docs assert that the state of the target's data directory
after rewind is equivalent to the source's data directory. But that
isn't true both because the base state is further back in time and
because the target's data directory will include the current state on
the source of any copied blocks.

In additional document clearly the state of the cluster after the
operation and also the operation sequencing dangers caused by copying
configuration files from the source.