Don't show worker info for sort node if no work done
In d2d8a229bc58a2014dce1c7a4fcdb6c5ab9fb8da we modified the
TuplesortMethod enum to be a bitmask. As such,
SORT_TYPE_STILL_IN_PROGRESS is now set to one. It'd previously been
explicitly set to 0, but with no comment, so that seemed reasonable.
However it seems that explain.c's show_sort_info method was using that 0
as a sentinel that the worker hadn't done any work, and thus it
shouldn't be printed (the worker's DSM is initialized to all zeros, so
this worked properly). Instead it should explicitly check for NULL.