pg_rewind: warn if we haven't checkpointed after promote
Because pg_rewind looks at the last checkpoint to determine the timeline
on the source server it can incorrectly report that the timelines are
the same if the source is a recently promoted standby that hasn't yet
completed a new checkpoint.

This is already documented, but we shouldn't tell the user a rewind
isn't necessary when it clearly is.