Helpful hint for qualified target column in UPDATE
Target columns in UPDATE ... SET must not be qualified with the target
table. The error message when a query contains this mistake is
unnecessarily confusing: "column "foo" of relation "foo" does not
exist". In the case where the column is qualified and the qualification
matches the table name (or alias) we can help the user by reminding them
of the syntax restriction for target columns.